refactor: migrate database connection and keyword caching to new structure
